refactor: add missing exports
