+++ src/__tests__/example-ipv4.ts
@@ -1,12 +1,4 @@
-import {
-  buildRegExp,
-  charRange,
-  choiceOf,
-  digit,
-  endOfString,
-  repeat,
-  startOfString,
-} from '../index';
+import { buildRegExp, charRange, choiceOf, digit, endOfString, repeat, startOfString } from '..';
 
 test('example: IPv4 address validator', () => {
   const octet = choiceOf(
